requirements.txt
art
rich
setuptools
telethon
PySocks
python-socks
faker
python-dotenv
asyncpg
cryptography
APScheduler
aiohttp
rarfile
SQLAlchemy
